Increase coverage
I already had tests that should have a bigger coverage than reported.
Apparently there is an [issue with
assert_cmd](https://github.com/assert-rs/assert_cmd/issues/9) that makes it not show
coverage correctly for tarpaulin.

So the workaround is to test as needed with assert_cmd and in the same
function call one that re-runs the command executing run_app and only
checks that the result is ok.