Do less things with travis
I don't need it to create a tag for every commit. Maybe in the future,
if it is a release branch, a tag could be created and the version number
be bumped or something like that. Not needed in a scenario where I am
the only tester. Or maybe only for the master branch.